Add benchmark script, update requirements and setup instructions
The commit introduces a number of important changes. First, it updates the `requirements.txt` file to include required dependencies like Numpy, Cython, Scikit-learn, and SciPy. This change allows users to know the required packages for the program to run successfully.

Furthermore, a new `benchmark_sklearn.py` script is added under the `benchmark` folder. The script includes various algorithms (like MeanShift, DBSCAN, MeanShiftPP and GridShiftPP) and their respective time performance on a sample dataset. This update will help measure and understand the efficiency of the algorithms implemented.

Also, the file setup.py is added to help in the installation of the Python package. The file includes a module for Cython to be built using the distutils extension enabling faster computations.

Lastly, the README file is updated to include the installation instructions for the package and the `__init__.py` files have been adjusted to ensure the right functions are exposed for import. All these changes collectively aim to increase the usability and performance of the ShiftClustering package.