Update clustering code for performance and usability changes
Updated clustering code to improve performance and usability. Enhancements include increasing the number of features (from 2 to 4) and centers (from 3 to 4,000) for the random dataset used for bench-marking. This will provide more realistic results and a stronger stress test for the clustering algorithms.

Renamed "gridshiftpp" to "gridshift" to better reflect the fact that it's an implementation of the Grid-Shift algorithm, without any specific enhancements like 'pp'.

10,000 samples are replaced with 50,000 samples for better benchmarking. Along with this, 2 features are replaced with 4 features and instead of 3, 4,000 centers are used in this commit.

To easier tie back the original work, acknowledgements have been added to the 'README.md' to credit the repo's that the project is based on.

Updated function benchmark_clustering to reflect the modifications.

Renamed GridShiftPP to GridShift class to reflect updated Source file name & re-appointed related calls.

These changes aim to provide a more accurate, effective and traceable clustering algorithm.