Update clustering parameters and remove unnecessary functions
The parameters of clustering algorithms were updated to achieve better clustering. The number of centers in the make_blobs function has been reduced to 1000. The bandwidth for all clustering algorithms was increased to 1.0 to better encapsulate the data nature. Also, the number of jobs for the MeanShift algorithm was set to 10 for faster execution and bin_seeding was set to True for better center initialization. In GridShift and MeanShiftPP the draw_box, update_hist, track, stitch_cy, and other unnecessary functions were removed. This significantly reduced code complexity without affecting functionality. Additionally, DBSCAN algorithm was commented out in benchmarking for unknown reasons. Instructions in the README file were also modified to reflect that the package is no longer under development and a reference to the original sources were included.