Update project version to 0.1.1 and adjust dependencies
This commit updates the project version from 0.1.0 to 0.1.1 in the `pyproject.toml` and `uv.lock` files. The Python version requirement has been modified to restrict it to versions 3.10 and below 3.13. Additionally, the `numpy` dependency has been updated to require version 1.26 or higher, while the `numba` and `torch` dependencies have been adjusted to require versions 0.60 and 2.7.0 respectively. These changes aim to enhance compatibility and ensure the package utilizes the latest features and improvements in its dependencies.